Extract process manager test modules to individual files
To support test reuse by external libs.
