+++ test/example_domain/money_transfer/transfer_money_command_handler.ex
@@ -1,12 +1,11 @@
 defmodule Commanded.ExampleDomain.TransferMoneyHandler do
   @moduledoc false
   alias Commanded.ExampleDomain.MoneyTransfer
-  alias MoneyTransfer.Commands.{TransferMoney}
+  alias MoneyTransfer.Commands.TransferMoney
 
   @behaviour Commanded.Commands.Handler
 
   def handle(%MoneyTransfer{} = aggregate, %TransferMoney{} = transfer_money) do
-    aggregate
-    |> MoneyTransfer.transfer_money(transfer_money)
+    MoneyTransfer.transfer_money(aggregate, transfer_money)
   end
 end
