+++ test/process_managers/process_manager_instance_test.exs
@@ -1,32 +1,17 @@
-defmodule Commanded.ProcessManager.ProcessManagerInstanceTest do
+defmodule Commanded.ProcessManagers.ProcessManagerInstanceTest do
   use Commanded.StorageCase
-  doctest Commanded.ProcessManagers.ProcessManagerInstance
 
-  alias Commanded.ExampleDomain.BankAccount
-  alias Commanded.ExampleDomain.BankAccount.Commands.WithdrawMoney
   alias Commanded.ExampleDomain.MoneyTransfer.Events.MoneyTransferRequested
   alias Commanded.ExampleDomain.TransferMoneyProcessManager
   alias Commanded.EventStore.RecordedEvent
-  alias Commanded.ProcessManagers.ProcessManagerInstance
-
-  defmodule NullHandler do
-    @behaviour Commanded.Commands.Handler
-
-    def handle(_aggregate, _command), do: []
-  end
-
-  defmodule Router do
-    use Commanded.Commands.Router
-
-    dispatch WithdrawMoney, to: NullHandler, aggregate: BankAccount, identity: :account_number
-  end
+  alias Commanded.ProcessManagers.{ProcessManagerInstance,NullRouter}
 
   test "process manager handles an event" do
     transfer_uuid = UUID.uuid4
     account1_uuid = UUID.uuid4
     account2_uuid = UUID.uuid4
 
-    {:ok, process_manager} = ProcessManagerInstance.start_link(Router, "TransferMoneyProcessManager", TransferMoneyProcessManager, transfer_uuid)
+    {:ok, process_manager} = ProcessManagerInstance.start_link(NullRouter, "TransferMoneyProcessManager", TransferMoneyProcessManager, transfer_uuid)
 
     event = %RecordedEvent{
       event_number: 1,
