+++ test/process_managers/resume_process_manager_test.exs
@@ -1,103 +1,20 @@
-defmodule Commanded.ProcessManager.ResumeProcessManagerTest do
+defmodule Commanded.ProcessManagers.ResumeProcessManagerTest do
   use Commanded.StorageCase
-
-  alias Commanded.Helpers.{ProcessHelper,Wait}
-  alias Commanded.ProcessManagers.{ProcessRouter,ProcessManagerInstance}
 
   import Commanded.Assertions.EventAssertions
 
-  defmodule ExampleAggregate do
-    defstruct [status: nil]
-
-    defmodule Commands do
-      defmodule StartProcess, do: defstruct [process_uuid: nil, status: nil]
-      defmodule ResumeProcess, do: defstruct [process_uuid: nil, status: nil]
-    end
-
-    defmodule Events do
-      defmodule ProcessStarted, do: defstruct [process_uuid: nil, status: nil]
-      defmodule ProcessResumed, do: defstruct [process_uuid: nil, status: nil]
-    end
-
-    alias Commands.{StartProcess,ResumeProcess}
-    alias Events.{ProcessStarted,ProcessResumed}
-
-    def start_process(%ExampleAggregate{}, %StartProcess{process_uuid: process_uuid, status: status}) do
-      %ProcessStarted{process_uuid: process_uuid, status: status}
-    end
-
-    def resume_process(%ExampleAggregate{}, %ResumeProcess{process_uuid: process_uuid, status: status}) do
-      %ProcessResumed{process_uuid: process_uuid, status: status}
-    end
-
-    # state mutatators
-
-    def apply(%ExampleAggregate{} = state, %ProcessStarted{status: status}), do: %ExampleAggregate{state | status: status}
-    def apply(%ExampleAggregate{} = state, %ProcessResumed{status: status}), do: %ExampleAggregate{state | status: status}
-  end
-
-  alias ExampleAggregate.Commands.{StartProcess,ResumeProcess}
-  alias ExampleAggregate.Events.{ProcessStarted,ProcessResumed}
-
-  defmodule CommandHandler do
-    @behaviour Commanded.Commands.Handler
-
-    def handle(%ExampleAggregate{} = aggregate, %StartProcess{} = start_process) do
-      aggregate
-      |> ExampleAggregate.start_process(start_process)
-    end
-
-    def handle(%ExampleAggregate{} = aggregate, %ResumeProcess{} = resume_process) do
-      aggregate
-      |> ExampleAggregate.resume_process(resume_process)
-    end
-  end
-
-  defmodule ExampleRouter do
-    use Commanded.Commands.Router
-
-    dispatch StartProcess, to: CommandHandler, aggregate: ExampleAggregate, identity: :process_uuid
-    dispatch ResumeProcess, to: CommandHandler, aggregate: ExampleAggregate, identity: :process_uuid
-  end
-
-  defmodule ExampleProcessManager do
-    use Commanded.ProcessManagers.ProcessManager,
-      name: "example-process-manager",
-      router: ExampleRouter
-
-    defstruct [
-      status_history: []
-    ]
-
-    alias ExampleAggregate.Events.ProcessStarted
-
-    def interested?(%ProcessStarted{process_uuid: process_uuid}), do: {:start, process_uuid}
-    def interested?(%ProcessResumed{process_uuid: process_uuid}), do: {:continue, process_uuid}
-
-    def handle(%ExampleProcessManager{}, %ProcessStarted{}), do: []
-    def handle(%ExampleProcessManager{}, %ProcessResumed{}), do: []
-
-    ## state mutators
-
-    def apply(%ExampleProcessManager{status_history: status_history} = process, %ProcessStarted{status: status}) do
-      %ExampleProcessManager{process |
-        status_history: status_history ++ [status]
-      }
-    end
-
-    def apply(%ExampleProcessManager{status_history: status_history} = process, %ProcessResumed{status: status}) do
-      %ExampleProcessManager{process |
-        status_history: status_history ++ [status]
-      }
-    end
-  end
+  alias Commanded.Helpers.{ProcessHelper,Wait}
+  alias Commanded.ProcessManagers.{ProcessRouter,ProcessManagerInstance}
+  alias Commanded.ProcessManagers.{ResumeRouter,ResumeProcessManager}
+  alias Commanded.ProcessManagers.ResumeAggregate.Commands.{StartProcess,ResumeProcess}
+  alias Commanded.ProcessManagers.ResumeAggregate.Events.{ProcessStarted,ProcessResumed}
 
   test "should resume a process manager with same state when process restarts" do
     process_uuid = UUID.uuid4
 
-    {:ok, process_router} = ExampleProcessManager.start_link()
+    {:ok, process_router} = ResumeProcessManager.start_link()
 
-    :ok = ExampleRouter.dispatch(%StartProcess{process_uuid: process_uuid, status: "start"})
+    :ok = ResumeRouter.dispatch(%StartProcess{process_uuid: process_uuid, status: "start"})
 
     assert_receive_event ProcessStarted, fn event ->
       assert event.process_uuid == process_uuid
@@ -117,9 +34,9 @@ defmodule Commanded.ProcessManager.ResumeProcessManagerTest do
     # wait for subscription to receive DOWN notification and remove subscription's PID
     :timer.sleep(1_000)
 
-    {:ok, process_router} = ExampleProcessManager.start_link()
+    {:ok, process_router} = ResumeProcessManager.start_link()
 
-    :ok = ExampleRouter.dispatch(%ResumeProcess{process_uuid: process_uuid, status: "resume"})
+    :ok = ResumeRouter.dispatch(%ResumeProcess{process_uuid: process_uuid, status: "resume"})
 
     wait_for_event(ProcessResumed, fn event -> event.process_uuid == process_uuid end)
