+++ test/process_managers/support/null_handler.ex
@@ -1,0 +1,6 @@
+defmodule Commanded.ProcessManagers.NullHandler do
+  @moduledoc false
+  @behaviour Commanded.Commands.Handler
+
+  def handle(_aggregate, _command), do: []
+end
