+++ test/process_managers/support/resume/resume_router.ex
@@ -1,0 +1,10 @@
+defmodule Commanded.ProcessManagers.ResumeRouter do
+  @moduledoc false
+  use Commanded.Commands.Router
+
+  alias Commanded.ProcessManagers.{ResumeAggregate,ResumeCommandHandler}
+  alias Commanded.ProcessManagers.ResumeAggregate.Commands.{StartProcess,ResumeProcess}
+
+  dispatch StartProcess, to: ResumeCommandHandler, aggregate: ResumeAggregate, identity: :process_uuid
+  dispatch ResumeProcess, to: ResumeCommandHandler, aggregate: ResumeAggregate, identity: :process_uuid
+end
